move header display and request signing
